+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed debugger display in `GameplayTagCountContainer`.
 - Fixed bug in `GameplayTagContainer.HasAll` that was returning false when the comparator container had only one tag.
 - Fixed bug where root game tags had an empty `HierarchyTags`.
+- Fixed `GameplayTagContainer.Clone()` implementation. (#1)
 
 ### Changed